add Profile

+++ database/migrations/2024_04_29_014639_create_promos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('title');
             $table->text('description');
             $table->text('image');
-            $table->integer('discount')->default(0);
+            $table->double('discount')->default(0);
             $table->integer('max_discount')->default(0);
             $table->integer('min_transaction')->default(0);
             $table->date('start_date');
